backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
groq
httpx
python-multipart
pydantic
GitPython
requests
reportlab
